Altered variables, commented out for loop
I have fixed the variables and echo commands at the top of the script so that they give the proper output when used with the example files. However, the script does not pass the run_tests.sh script (I assume because it is not dynamic). I commented out the for loop in the script because it was not giving me proper output and I started working on a new for loop at the bottom of the script, but it seems to be having some issues as well, so I commented it out as well.