Fixed total count variable.
The existing code had a very complex conditional branch in the loop which created an intermediate file.
I commented it out and replaced it with a single variable declared before the loop, set to "0".
In each iteration of the loop, the value of "$number" of sequences in the "$var" .fasta file is added to the running total.
It looks like this: ct=`expr $ct + $number`